Add autoscaling of images to mobile browsers
Currently, on mobile devices that are not wide enough for the images
in the blog, the image sets the max width of the blog and downscales
all the text relative to it, leading to crushed text and a large empty
space on the right side.

I've reproduced this using Google Chrome developer tools, setting the
device mode to emulate an iPhone X, but it appears to occur any time
the mobile device has a screen width smaller than the image
to be displayed.

The code change here hints that the max width of the image
should be 100% of the screen width, not an absolute number of pixels,
and the height will be automatically adjusted to maintain
the image aspect ratio.